sertis_api/requirements.txt
awswrangler==2.2.0
beautifulsoup4==4.9.3
boto3==1.16.48
botocore==1.19.48
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
curlify==2.2.1
DateTime==4.3
Django==2.2.3
django-cors-headers==3.1.0
django-extensions==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
facebook-business==9.0.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
idna==2.8
jmespath==0.9.4
lxml==4.6.2
Markdown==3.1.1
mock==4.0.2
mysqlclient==1.4.6
numpy==1.19.4
oauth2client==4.1.3
pandas==1.1.5
pg8000==1.16.6
protobuf==3.9.0
pyarrow==2.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycountry==20.7.3
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.0
pytz==2020.5
redshift-connector==2.0.872
requests==2.25.1
rsa==4.0
s3transfer==0.3.3
scramp==1.2.0
six==1.12.0
soupsieve==2.1
SQLAlchemy==1.2.19
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.26.2
zope.interface==4.6.0
